backend/requirements.txt
fastapi
uvicorn
google-cloud-speech
pydub
python-multipart
